Fix: Create AddImage to be re-useable
